+++ backend/src/services/simulation-service/team-simulator/member-state.test.ts
@@ -1,6 +1,7 @@
 import { mocks } from '@src/bun/index.js';
 import { calculateFrequencyWithEnergy } from '@src/services/calculator/help/help-calculator.js';
-import { CookingState } from '@src/services/simulation-service/team-simulator/cooking-state.js';
+import { CookingState } from '@src/services/simulation-service/team-simulator/cooking-state/cooking-state.js';
+import { defaultUserRecipes } from '@src/services/simulation-service/team-simulator/cooking-state/cooking-utils.js';
 import { MemberState } from '@src/services/simulation-service/team-simulator/member-state.js';
 import { TeamSimulatorUtils } from '@src/services/simulation-service/team-simulator/team-simulator-utils.js';
 import { TimeUtils } from '@src/utils/time-utils/time-utils.js';
@@ -69,7 +70,7 @@ const settings: TeamSettingsExt = {
   stockpiledIngredients: emptyIngredientInventoryFloat()
 };
 
-const cookingState: CookingState = new CookingState(settings);
+const cookingState: CookingState = new CookingState(settings, defaultUserRecipes());
 
 describe('results', () => {
   it('should return correct results after multiple iterations', () => {