+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [1.35.0](https://github.com/nerolis-lab/nerolis-lab/compare/v1.34.0...v1.35.0) (2025-02-19)
+
+
+### Bug Fixes
+
+* ingredient selection updates on update only ([8b270e8](https://github.com/nerolis-lab/nerolis-lab/commit/8b270e8be4e1c8b71c3ec61387e8e0a14b52b90f))
+* recipe mobile tabe width on safari works ([45b3405](https://github.com/nerolis-lab/nerolis-lab/commit/45b34056e7aeb2a34927a2faaf89a71739a7ab96))
+* wraps sort button on tiny mobile ([d2d76ac](https://github.com/nerolis-lab/nerolis-lab/commit/d2d76ac78b1d6bac0f423667597088d15dbcd87d))
+
+
+### Features
+
+* adds asc/desc sort support ([707dd40](https://github.com/nerolis-lab/nerolis-lab/commit/707dd40984f2ffbfe4cb5edece003601bec6d6d1))
+
+# [1.34.0](https://github.com/nerolis-lab/nerolis-lab/compare/v1.33.0...v1.34.0) (2025-02-19)
+
+
+### Features
+
+* recipe page and user recipe level ([c294496](https://github.com/nerolis-lab/nerolis-lab/commit/c2944961431e12b42f457040c7147b89cb838818))
+
+# [1.33.0](https://github.com/nerolis-lab/nerolis-lab/compare/v1.32.0...v1.33.0) (2025-02-15)
+
+
+### Features
+
+* dynamic max ingredient bonuses instead of static ones ([b1d2cb0](https://github.com/nerolis-lab/nerolis-lab/commit/b1d2cb02ef075fc40070c75e668388a0a232eb2f))
+
 # [1.32.0](https://github.com/nerolis-lab/nerolis-lab/compare/v1.31.1...v1.32.0) (2025-02-11)